Sequence Regression Strategy

+++ trading-strategies-impl/bachelor-strategies-impl/src/main/java/com/becker/freelance/strategies/regression/sequence/shared/LogReturnInverseTransformer.java
@@ -1,16 +1,15 @@
 package com.becker.freelance.strategies.regression.sequence.shared;
 
 import com.becker.freelance.commons.timeseries.TimeSeriesEntry;
-import com.becker.freelance.math.Decimal;
 
 import java.util.function.Function;
 
-public interface LogReturnInverseTransformer extends Function<TimeSeriesEntry, Decimal[]> {
+public interface LogReturnInverseTransformer extends Function<TimeSeriesEntry, Double[]> {
 
-    public Decimal[] transformLogReturnsToPrice(TimeSeriesEntry initialPrice);
+    public Double[] transformLogReturnsToPrice(TimeSeriesEntry initialPrice);
 
     @Override
-    default Decimal[] apply(TimeSeriesEntry initialPrice) {
+    default Double[] apply(TimeSeriesEntry initialPrice) {
         return transformLogReturnsToPrice(initialPrice);
     }
 }
